Use uint64_t for Timer{Micros,Millis}
GitOrigin-RevId: ac6efe14557bf810e0432703eb2c4936028e9d49

+++ arduino/cores/arduino/wiring.cpp
@@ -36,9 +36,9 @@ void delayMicroseconds(unsigned int us) {
 }
 
 unsigned long millis() {
-    return coralmicro::TimerMillis();
+    return static_cast<unsigned long>(coralmicro::TimerMillis());
 }
 
 unsigned long micros() {
-    return coralmicro::TimerMicros();
+    return static_cast<unsigned long>(coralmicro::TimerMicros());
 }
